% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gpsa_collapse_duplicates_maxabs}
\alias{gpsa_collapse_duplicates_maxabs}
\title{Collapse duplicate gene names by max absolute value}
\usage{
gpsa_collapse_duplicates_maxabs(stats)
}
\arguments{
\item{stats}{Named numeric vector}
}
\value{
Named numeric vector with unique names
}
\description{
Collapse duplicate gene names by max absolute value
}
