% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{gpsa_compile_query}
\alias{gpsa_compile_query}
\title{Compile gene-set query}
\usage{
gpsa_compile_query(db, query_sets, min_size = 20)
}
\arguments{
\item{db}{A gpsa_db object}

\item{query_sets}{Named list of gene sets, each with $genes (character vector),
$sign (+1 or -1), and optional $weight (default 1)}

\item{min_size}{Minimum genes after intersection with DB (default: 20)}
}
\value{
Compiled query object
}
\description{
Compiles one or more gene sets into query weights for search.
}
\examples{
\dontrun{
db <- gpsa_load_db()
query <- list(
  Up = list(genes = up_genes, sign = +1),
  Down = list(genes = down_genes, sign = -1)
)
cq <- gpsa_compile_query(db, query)
}
}
