% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{gpsa_compile_signature_vector}
\alias{gpsa_compile_signature_vector}
\title{Compile full geneList to vector query weights}
\usage{
gpsa_compile_signature_vector(
  db,
  geneList,
  transform = c("rankZ", "zscore"),
  ties_method = "average",
  normalize = TRUE,
  min_size = 200
)
}
\arguments{
\item{db}{A gpsa_db object}

\item{geneList}{Named numeric vector (gene symbols as names)}

\item{transform}{Transformation: "rankZ" or "zscore"}

\item{ties_method}{Ties method for ranking}

\item{normalize}{L2-normalize weights?}

\item{min_size}{Minimum overlap with DB}
}
\value{
Compiled vector query object
}
\description{
Converts a full geneList (named numeric vector) into query weights
using rankZ transformation.
}
