% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gpsa_fgsea_updown}
\alias{gpsa_fgsea_updown}
\title{GSEA-style validation with fgsea}
\usage{
gpsa_fgsea_updown(
  geneList,
  up_genes,
  down_genes,
  nperm = 2000,
  minSize = 10,
  maxSize = 5000
)
}
\arguments{
\item{geneList}{Named numeric vector}

\item{up_genes}{Up-regulated gene set}

\item{down_genes}{Down-regulated gene set}

\item{nperm}{Number of permutations}

\item{minSize}{Minimum gene set size}

\item{maxSize}{Maximum gene set size}
}
\value{
fgsea result data.frame
}
\description{
Runs fgsea on up and down gene sets for validation.
}
