% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gpsa_gsea_running}
\alias{gpsa_gsea_running}
\title{Compute GSEA running enrichment score}
\usage{
gpsa_gsea_running(geneList, geneSet, gsea_param = 1)
}
\arguments{
\item{geneList}{Named numeric vector (sorted descending)}

\item{geneSet}{Character vector of gene symbols}

\item{gsea_param}{GSEA weight parameter (default: 1)}
}
\value{
List with ES, peak, running scores, hit positions
}
\description{
Compute GSEA running enrichment score
}
