% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{gpsa_install_deps}
\alias{gpsa_install_deps}
\title{Install GPSA dependencies}
\usage{
gpsa_install_deps(
  install_minimal = TRUE,
  install_optional = TRUE,
  ask = FALSE,
  update = FALSE
)
}
\arguments{
\item{install_minimal}{Install minimal dependencies (rhdf5, Matrix)?}

\item{install_optional}{Install optional packages (matrixStats, msigdbr, fgsea, ggplot2)?}

\item{ask}{Ask before installing?}

\item{update}{Update existing packages?}
}
\value{
TRUE invisibly
}
\description{
Helper to install required and optional packages.
}
