% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gpsa_pick_signature}
\alias{gpsa_pick_signature}
\title{Pick a signature by gene and cell line}
\usage{
gpsa_pick_signature(db, pbgene, cellline = NULL, fallback = NULL)
}
\arguments{
\item{db}{A gpsa_db object}

\item{pbgene}{Perturbation gene name}

\item{cellline}{Cell line name (optional)}

\item{fallback}{Fallback signature ID if not found}
}
\value{
Signature ID or NULL
}
\description{
Pick a signature by gene and cell line
}
