% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gpsa_prepare_geneList}
\alias{gpsa_prepare_geneList}
\title{Prepare a geneList for GSEA/search}
\usage{
gpsa_prepare_geneList(
  x,
  decreasing = TRUE,
  collapse_duplicates = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{x}{Named numeric vector (gene symbols as names)}

\item{decreasing}{Sort decreasing?}

\item{collapse_duplicates}{Collapse duplicates by max|value|?}

\item{na.rm}{Remove NA values?}
}
\value{
Sorted named numeric vector
}
\description{
Prepare a geneList for GSEA/search
}
