% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gpsa_rank_in}
\alias{gpsa_rank_in}
\title{Get rank of a signature in a result data.frame}
\usage{
gpsa_rank_in(df, signature_id)
}
\arguments{
\item{df}{Result data.frame with signature_id column}

\item{signature_id}{Signature ID to find}
}
\value{
Integer rank (row number) or NA
}
\description{
Get rank of a signature in a result data.frame
}
