% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gpsa_score_stats}
\alias{gpsa_score_stats}
\title{Compute score statistics}
\usage{
gpsa_score_stats(score_vec, robust = TRUE)
}
\arguments{
\item{score_vec}{Numeric vector of scores}

\item{robust}{Use robust statistics (median/MAD)?}
}
\value{
List with tau, p_rank, z_ref, p_ref
}
\description{
Compute score statistics
}
