% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gpsa_search}
\alias{gpsa_search}
\title{Search with gene-set query}
\usage{
gpsa_search(
  db,
  query_sets,
  return_components = TRUE,
  block_s = NULL,
  robust = TRUE
)
}
\arguments{
\item{db}{A gpsa_db object}

\item{query_sets}{Named list of gene sets (see gpsa_compile_query)}

\item{return_components}{Return per-channel NES scores?}

\item{block_s}{Block size for reading (NULL = auto)}

\item{robust}{Use robust statistics?}
}
\value{
List with res (data.frame), compile info, and optionally NES_raw/NES_signed
}
\description{
Performs similarity search using one or more gene sets.
}
\examples{
\dontrun{
db <- gpsa_load_db(preload_Z = TRUE)

# Single gene set query
query <- list(MySet = list(genes = my_genes, sign = +1))
out <- gpsa_search(db, query)
head(out$res)

# Up/Down query
query <- list(
  Up = list(genes = up_genes, sign = +1),
  Down = list(genes = down_genes, sign = -1)
)
out <- gpsa_search(db, query, return_components = TRUE)
}
}
