% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gpsa_set_h5path}
\alias{gpsa_set_h5path}
\title{Set GPSA H5 file path}
\usage{
gpsa_set_h5path(h5file)
}
\arguments{
\item{h5file}{Path to HDF5 file}
}
\value{
Previous value (invisibly)
}
\description{
Convenience function to set the default HDF5 file path.
}
