% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{print.gpsa_db}
\alias{print.gpsa_db}
\title{Print method for gpsa_db}
\usage{
\method{print}{gpsa_db}(x, ...)
}
\arguments{
\item{x}{A gpsa_db object}

\item{...}{Ignored}
}
\description{
Print method for gpsa_db
}
